Use test fixture for global setup of integration tests
